docs(readme): rename example variable `file` to `run`
Overloading the name `file` to refer to both an internal variable and
the file itself can lead to confusion when reading the example syntax.
By renaming the variable to `run`, it should be more clear that the
first two examples refer to a variable, while in the third `[name]`
pulls the filename.

Signed-off-by: Zac Delventhal <delventhalz@gmail.com>